// viewer/src/common/types.ts
// DO NOT edit this file, edit src/common in the root of this repo instead

import { JsonObject } from "./json";

export interface Config extends JsonObject {}

export interface E3Record extends JsonObject {
  caseNumber: string;
  caseStatus: string;
  receivedDate: string;
  decisionDate: string;
  originalCertDate: string;
  visaClass: string;
  jobTitle: string;
  socCode: string;
  socTitle: string;
  fullTimePosition: string;
  beginDate: string;
  endDate: string;
  totalWorkerPositions: string;
  newEmployment: string;
  continuedEmployment: string;
  changePreviousEmployment: string;
  newConcurrentEmployment: string;
  changeEmployer: string;
  amendedPetition: string;
  employerName: string;
  tradeNameDba: string;
  employerAddress1: string;
  employerAddress2: string;
  employerCity: string;
  employerState: string;
  employerPostalCode: string;
  employerCountry: string;
  employerProvince: string;
  employerPhone: string;
  employerPhoneExt: string;
  naicsCode: string;
  employerPocLastName: string;
  employerPocFirstName: string;
  employerPocMiddleName: string;
  employerPocJobTitle: string;
  employerPocAddress_1: string;
  employerPocAddress_2: string;
  employerPocCity: string;
  employerPocState: string;
  employerPocPostalCode: string;
  employerPocCountry: string;
  employerPocProvince: string;
  employerPocPhone: string;
  employerPocPhoneExt: string;
  employerPocEmail: string;
  agentRepresentingEmployer: string;
  agentAttorneyLastName: string;
  agentAttorneyFirstName: string;
  agentAttorneyMiddleName: string;
  agentAttorneyAddress1: string;
  agentAttorneyAddress2: string;
  agentAttorneyCity: string;
  agentAttorneyState: string;
  agentAttorneyPostalCode: string;
  agentAttorneyCountry: string;
  agentAttorneyProvince: string;
  agentAttorneyPhone: string;
  agentAttorneyPhoneExt: string;
  agentAttorneyEmailAddress: string;
  lawfirmNameBusinessName: string;
  stateOfHighestCourt: string;
  nameOfHighestStateCourt: string;
  worksiteWorkers: string;
  secondaryEntity: string;
  secondaryEntityBusinessName: string;
  worksiteAddress1: string;
  worksiteAddress2: string;
  worksiteCity: string;
  worksiteCounty: string;
  worksiteState: string;
  worksitePostalCode: string;
  wageRateOfPayFrom: string;
  wageRateOfPayTo: string;
  wageUnitOfPay: string;
  prevailingWage: string;
  pwUnitOfPay: string;
  pwTrackingNumber: string;
  pwWageLevel: string;
  pwOesYear: string;
  pwOtherSource: string;
  pwOtherYear: string;
  pwSurveyPublisher: string;
  pwSurveyName: string;
  totalWorksiteLocations: string;
  agreeToLcStatement: string;
  h1BDependent: string;
  willfulViolator: string;
  supportH1B: string;
  statutoryBasis: string;
  appendixAAttached: string;
  publicDisclosure: string;
  preparerLastName: string;
  preparerFirstName: string;
  preparerMiddleInitial: string;
  preparerBusinessName: string;
  preparerEmail: string;
}

export interface SearchArgs {
  employerName?: string;
  jobTitle?: string;
  worksiteCity?: string;
  worksiteState?: string;
  decisionDate?: string;
}